Create L10Q1_dice
Lab 10 Problem 1: Prompt the user for the number of dice they would like to roll. Generate a random
number from a six-sided die (assume they roll one at a time) and print that number.
Do this for each roll. Get the sum of all the dice rolls and print it out.